backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
cachetools==5.5.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
